Fix passing wrong types for parameters due to copy and paste classes

+++ src/main/java/com/example/javacms/service/ArticleService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface ArticleService {
     List<Article> findAll();
 
-    Article findById(int id);
+    Article findById(long id);
 
     void save(Article article);
 
     void update(Article article);
 
-    void deleteById(int id);
+    void deleteById(long id);
 }
