+++ src/main/java/com/example/javacms/service/CommentService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface CommentService {
     List<Comment> findAll();
 
-    Comment findById(int id);
+    Comment findById(long id);
 
     void save(Comment comment);
 
     void update(Comment comment);
 
-    void deleteById(int id);
+    void deleteById(long id);
 }
